+++ conda_pypi/plugin.py
@@ -7,7 +7,7 @@ from conda_pypi import cli
 from conda_pypi import post_command
 from conda_pypi.main import ensure_target_env_has_externally_managed
 from conda.plugins.types import CondaPackageExtractor
-from conda_pypi.package_extractor.extract_whl import extract_whl_as_conda_pkg
+from conda_pypi.package_extractors.whl import extract_whl_as_conda_pkg
 
 
 @hookimpl
